Split "render" module into separate "renderer" and "render" modules
