+++ packages/blitz/src/lib.rs
@@ -7,9 +7,9 @@ lib.rs -> DioxusNative -> Window -> Document
 mod devtools;
 mod fontcache;
 mod imagecache;
-pub mod render;
+pub mod renderer;
 mod util;
 mod viewport;
 
-pub use render::*;
+pub use renderer::*;
 pub use viewport::Viewport;
