+++ packages/blitz/src/renderer/multicolor_rounded_rect.rs
